Add files via upload
Here we solve the classification problem using the kernel trick in the dual form using the same training samples. The objective here would to get calculate the kernel function instead of determining the feature representation. Then the dot product of training samples can be switched with the kernel functions. Then the dual form is optimized for alpha.
The b can be calculated using this alpha. The f(x) = (W*t * Q(x) + b) can also be calculated from kernel function and alpha.