Add files via upload
Here the classification is done using KNN classifier. The value for K would be 3 and 5. The metric for distance between two points would be L2 norm. For calculating the matrix with distance of training samples for each other , instead of creating a matrix of 2000 X 2000, I have used the fact that the distance from i to j would be equal to distance from j to i and hence used a matrix of dimensions (2000 X 1999)/2 to calculate the distance of the training samples.